Core: Implemented fixed ratio port constraints.
